+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Fixed bugs with item getting duplicated on pick up after kicking.
+- Smoothed out item physics.
 
 ## [1.2.2] - 2026-04-08